Rename FactoryFunctionBuilder to ConstructorBuilder
